mobile/e2e/requirements.txt
appium-python-client==1.0.2
pytest==6.0.1
